app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
cryptography>=41.0.0
alembic==1.12.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
jinja2==3.1.2
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf>=4.25.0
aiokafka==0.10.0
